created utility for easily adding vite plugins
