+++ packages/start/src/config/defineConfig.ts
@@ -2,16 +2,17 @@ import type { SolidifrontConfig } from "./types";
 
 import defu from "defu";
 import path from "path";
+import generateShopifyLocalesPlugin from "@solidifront/vite-plugin-generate-shopify-locales";
+
 import {
   type SolidStartInlineConfig,
   defineConfig as defineSolidConfig,
 } from "@solidjs/start/config";
-
 import { Project } from "ts-morph";
 
-import generateShopifyLocalesPlugin from "@solidifront/vite-plugin-generate-shopify-locales";
 import { handleMiddleware } from "./utils.js";
 import { solidifrontMiddlewareSetup } from "./plugins/index.js";
+import { attachPlugins } from "./viteHelpers";
 
 export namespace defineConfig {
   export type Config = SolidifrontConfig;
@@ -45,48 +46,19 @@ export function defineConfig(baseConfig: defineConfig.Config = {}) {
     handleMiddleware(project, config.middleware!);
   }
 
-  if (typeof vite === "function") {
-    const defaultVite = vite;
-
-    vite = (options) => {
-      const viteConfig = defaultVite(options);
-      return defu(viteConfig, {
-        plugins: [solidifrontMiddlewareSetup(project, config.solidifront)],
-      });
-    };
-  } else if (typeof vite === "object") {
-    vite.plugins = (vite.plugins || []).concat([
-      solidifrontMiddlewareSetup(project, config.solidifront),
-    ]);
-  }
+  vite = attachPlugins(vite, [
+    solidifrontMiddlewareSetup(project, config.solidifront),
+  ]);
 
   if (needsLocalization) {
-    if (typeof vite === "function") {
-      const defaultVite = vite;
-
-      vite = (options) => {
-        const viteConfig = defaultVite(options);
-        return defu(viteConfig, {
-          plugins: [
-            generateShopifyLocalesPlugin({
-              defaultLocale:
-                config.solidifront?.localization?.defaultLocale ?? undefined,
-              debug: process.env.NODE_ENV === "development",
-              namespace: "@solidifront/start/locales",
-            }),
-          ],
-        });
-      };
-    } else if (typeof vite === "object") {
-      vite.plugins = (vite.plugins || []).concat([
-        generateShopifyLocalesPlugin({
-          defaultLocale:
-            config.solidifront?.localization?.defaultLocale ?? undefined,
-          debug: process.env.NODE_ENV === "development",
-          namespace: "@solidifront/start/locales",
-        }),
-      ]);
-    }
+    vite = attachPlugins(vite, [
+      generateShopifyLocalesPlugin({
+        defaultLocale:
+          config.solidifront?.localization?.defaultLocale ?? undefined,
+        debug: process.env.NODE_ENV === "development",
+        namespace: "@solidifront/start/locales",
+      }),
+    ]);
   }
 
   return defineSolidConfig({
