+++ packages/start/src/config/viteHelpers/index.ts
@@ -1,0 +1,1 @@
+export { attachPlugins } from "./attachPlugins.js";
